refactor(py/web/servers): add multi-coroutine server manager
CHANGELOG:
- [ ] Move the multi-coroutine server manager implementation
      from the aioia package into genkit.
- [ ] Clean up the sample and rename it from
      aioia_multi_server -> multi_server.
