+++ py/packages/genkit/src/genkit/web/servers/__init__.py
@@ -3,7 +3,12 @@
 
 """Run multiple servers in a single process."""
 
-from ._adapters import GranianAdapter, UvicornAdapter
+from ._adapters import (
+    ASGIServerAdapter,
+    GranianAdapter,
+    ServerType,
+    UvicornAdapter,
+)
 from ._base_server import AbstractBaseServer
 from ._info import get_health_info, get_server_info
 from ._loop import run_loop
@@ -11,15 +16,16 @@ from ._managers import ServersManager
 from ._server import Server, ServerConfig, ServerLifecycle
 
 __all__ = [
+    'ASGIServerAdapter',
     'AbstractBaseServer',
     'GranianAdapter',
     'Server',
     'ServerConfig',
     'ServerLifecycle',
+    'ServerType',
     'ServersManager',
     'UvicornAdapter',
     'get_health_info',
     'get_server_info',
     'run_loop',
-    'run_servers',
 ]
