+++ py/samples/multi_server/pyproject.toml
@@ -1,14 +1,15 @@
 [project]
 dependencies = [
   "aioia",
+  "genkit",
   "asgiref>=3.8.1",
   "litestar>=2.15.1",
   "starlette>=0.46.1",
   "structlog>=25.2.0",
   "uvloop>=0.21.0",
 ]
-description = "Sample implementation to exercise the aioia multi server manager."
-name = "aioia-multi-server"
+description = "Sample implementation to exercise the Genkit multi server manager."
+name = "multi-server"
 readme = "README.md"
 requires-python = ">=3.12"
 version = "0.1.0"
